refactor(audiobook): consolidate duplicated time formatter
